+++ MapUtil/src/test/java/evolution/util/UpdateCountTest.java
@@ -8,7 +8,7 @@ import org.junit.Test;
 public class UpdateCountTest {
 	@Test
 	public void testUpdateCount4SingleKeyCase() {
-		Map<String, Integer> map = new HashMap<>();
+		Map<String, Double> map = new HashMap<>();
 		MapUtil.updateCount("apple", map);
 		MapUtil.updateCount("pear", map);
 		MapUtil.updateCount("apple", map);
@@ -17,13 +17,13 @@ public class UpdateCountTest {
 	
 	@Test
 	public void testUpdateCount4DoubleMapCase() {
-		Map<String, Integer> map0 = new HashMap<>();
-		map0.put("apple", 3);
-		map0.put("pear", 4);
-		Map<String, Integer> map1 = new HashMap<>();
-		map1.put("apple", 5);
-		map1.put("pear", 2);
-		Map<String, Integer> result = MapUtil.updateCount(map0, map1);
+		Map<String, Double> map0 = new HashMap<>();
+		map0.put("apple", 3d);
+		map0.put("pear", 4d);
+		Map<String, Double> map1 = new HashMap<>();
+		map1.put("apple", 5d);
+		map1.put("pear", 2d);
+		Map<String, Double> result = MapUtil.updateCount(map0, map1);
 		System.out.println(result);
 	}
 }
